+++ tests/test_redis_throttled_queue.py
@@ -35,6 +35,40 @@ def test_simple(redis_conn: StrictRedis, redis_monitor):
     assert len(queue) == 0
 
 
+def test_cleanup(redis_conn: StrictRedis, redis_monitor):
+    queue = ThrottledQueue(redis_conn, "test", limit=5, resolution=Resolution.SECOND)
+    for pos, item in enumerate(range(10)):
+        queue.push('aaaaaa', f'a{item}', priority=10 - pos)
+    for pos, item in enumerate(range(10)):
+        queue.push('bbbbbb', f'b{item}', priority=10 - pos)
+
+    assert len(queue) == 20
+    items = ','.join(queue.pop() for _ in range(10))
+    assert items == 'a0,b0,a1,b1,a2,b2,a3,b3,a4,b4'
+    assert len(queue) == 10
+    assert queue.pop() is None
+    assert len(queue) == 10
+
+    sleep(1)
+
+    queue.cleanup()
+    assert len(queue) == 0
+    assert queue.pop() is None
+
+
+def test_cleanup_directly(redis_conn: StrictRedis, redis_monitor):
+    queue = ThrottledQueue(redis_conn, "test", limit=5, resolution=Resolution.SECOND)
+    for pos, item in enumerate(range(10)):
+        queue.push('aaaaaa', f'a{item}', priority=10 - pos)
+    for pos, item in enumerate(range(10)):
+        queue.push('bbbbbb', f'b{item}', priority=10 - pos)
+
+    assert len(queue) == 20
+    queue.cleanup()
+    assert len(queue) == 0
+    assert queue.pop() is None
+
+
 def test_priority(redis_conn: StrictRedis, redis_monitor):
     queue = ThrottledQueue(redis_conn, "test", limit=5, resolution=Resolution.SECOND)
     for pos, item in enumerate(range(10)):
@@ -50,6 +84,7 @@ def test_priority(redis_conn: StrictRedis, redis_monitor):
     assert len(queue) == 10
 
     sleep(1)
+    assert queue.idle_seconds == pytest.approx(1, 0.01)
 
     items = ','.join(queue.pop() for _ in range(10))
     assert items == 'a5,b4,a6,b3,a7,b2,a8,b1,a9,b0'
@@ -78,12 +113,14 @@ def test_extras(redis_conn: StrictRedis, redis_monitor):
     assert items in ['b0,aX,b1,a1,b2,a2,b3,a3,b4,c0,c1,None,None']
 
     sleep(1)
+    assert queue.idle_seconds == pytest.approx(1, 0.01)
 
     items = ','.join(str(queue.pop()) for _ in range(12))
     assert len(queue) == 2
     assert items in ['a4,b5,a5,b6,a6,b7,a7,b8,a8,b9,None,None']
 
     sleep(1)
+    assert queue.idle_seconds == pytest.approx(1, 0.01)
 
     items = ','.join(str(queue.pop()) for _ in range(4))
     assert len(queue) == 0